+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - bump Android Gradle Plugin from _8.3.2_ to _8.4.0_ (compile-only dependency) in https://github.com/fluxo-kt/fluxo-kmp-conf/pull/48.
 - bump [io.nlopez.compose.rules:detekt](https://github.com/mrmans0n/compose-rules) from _0.3.15_ to _0.3.20_ in https://github.com/fluxo-kt/fluxo-kmp-conf/pull/53.
 - bump [binary-compatibility-validator-js](https://github.com/fluxo-kt/fluxo-bcv-js) from _0.2.0_ to _0.3.0_ in https://github.com/fluxo-kt/fluxo-kmp-conf/pull/52.
+- bump [proguard-core](https://github.com/Guardsquare/proguard-core) from _9.1.2_ to _9.1.3_ in https://github.com/fluxo-kt/fluxo-kmp-conf/pull/49.
 
 
 ## [0.8.0] - 2024-04-22